refactor the main/renderer code so it is more clear where files land
