+++ src/main/menu.js
@@ -1,8 +1,8 @@
 var dialog = require('electron').dialog
 
-var userConfig = require('./lib/user-config')
-var exportData = require('./lib/export-data')
-var i18n = require('./lib/i18n')
+var userConfig = require('./user-config')
+var exportData = require('./export-data')
+var i18n = require('../i18n')
 
 module.exports = function (app) {
   var template = [
