+++ src/main/osm-stats.js
