+++ src/main/polygon-feature.js
