+++ src/renderer/app.js
