+++ src/renderer/components/ConvertButton.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Paper from '@material-ui/core/Paper'
 import Button from '@material-ui/core/Button'
-import i18n from '../lib/i18n'
+import i18n from '../../i18n'
 
 const ConvertButton = function ({ onClick, features }) {
   function click () {
