+++ src/renderer/components/MapFilter.js
