+++ src/renderer/components/Overlay.js
