+++ src/renderer/components/ProgressBar.js
