+++ src/renderer/components/View.js
