+++ src/renderer/components/dialogs/Error.js
